requirements.txt
fastapi~=0.70.0
uvicorn~=0.15.0
pydantic~=1.8.2
starlette~=0.16.0
httpx~=0.20.0
python-decouple
